+++ apps/web/src/stores/themeStore.ts
@@ -1,0 +1,46 @@
+import { create } from 'zustand';
+
+type Theme = 'light' | 'dark';
+
+interface ThemeStore {
+  theme: Theme;
+  toggleTheme: () => void;
+  setTheme: (theme: Theme) => void;
+}
+
+const getInitialTheme = (): Theme => {
+  const savedTheme = localStorage.getItem('theme') as Theme | null;
+  if (savedTheme) {
+    return savedTheme;
+  }
+  if (window.matchMedia('(prefers-color-scheme: dark)').matches) {
+    return 'dark';
+  }
+  return 'light';
+};
+
+export const useThemeStore = create<ThemeStore>((set) => ({
+  theme: getInitialTheme(),
+  toggleTheme: () =>
+    set((state) => {
+      const newTheme = state.theme === 'light' ? 'dark' : 'light';
+      updateDomAndStorage(newTheme);
+      return { theme: newTheme };
+    }),
+  setTheme: (theme) => {
+    updateDomAndStorage(theme);
+    set({ theme });
+  },
+}));
+
+const updateDomAndStorage = (theme: Theme) => {
+  const root = window.document.documentElement;
+  if (theme === 'dark') {
+    root.classList.add('dark');
+  } else {
+    root.classList.remove('dark');
+  }
+  localStorage.setItem('theme', theme);
+};
+
+updateDomAndStorage(getInitialTheme());
